Add a range resolution to fix issues with venturocket/slider
Generally we want to avoid resolutions in libraries, but this range
resolution matches the required specified angular dependency range
but also resolves the fact that venturocket slider has not been
updated and still says it requires angular 1.2.x when it works with
later versions.
This range resolution will also need to be included in projects which
use this library package.